Reduce data type widths where possible

+++ src/koala/example_graphs.py
@@ -614,12 +614,12 @@ def make_amorphous(L, return_points = False):
     points = np.random.uniform(size=(L**2,2))
     lattice = voronization.generate_lattice(points)
     coloring = graph_color.color_lattice(lattice)
-    gs_flux_sector = np.array([ground_state_ansatz(p.n_sides) for p in lattice.plaquettes])
+    gs_flux_sector = np.array([ground_state_ansatz(p.n_sides) for p in lattice.plaquettes], dtype = np.int8)
     ujk = flux_finder.find_flux_sector(lattice, gs_flux_sector)
     if return_points: return lattice, coloring, ujk, return_points
     return lattice, coloring, ujk
 
 def make_honeycomb(L):
     lattice, coloring = generate_honeycomb(L, return_coloring = True)
-    ujk = np.ones(lattice.n_edges)
-    return lattice, coloring, ujk
+    ujk = np.ones(lattice.n_edges, dtype = np.int8)
+    return lattice, coloring.astype(np.int8), ujk
